Day4(Exceptions) - Done!

+++ src/main/java/com/jschool/Day4_Exceptions/ClientAccount.java
@@ -3,25 +3,25 @@ package com.jschool.Day4_Exceptions;
 import com.jschool.Day4_Exceptions.MyExceptions.AccountException.AccountBalanceException;
 
 public class ClientAccount {
-    private long balance;
+    private int balance;
 
     public ClientAccount() {
         this.balance = 0;
     }
 
-    public long getBalance() {
+    public int getBalance() {
         return balance;
     }
 
-    private void setBalance(long balance) {
+    private void setBalance(int balance) {
         this.balance = balance;
     }
 
-    public void plusBalance(long sum) {
+    public void plusBalance(int sum) {
         setBalance(getBalance() + sum);
     }
 
-    public void minusBalance(long sum) throws AccountBalanceException {
+    public void minusBalance(int sum) throws AccountBalanceException {
 
         if (getBalance() - sum < 0) {
             throw new AccountBalanceException("Недостаточно средств для выполнения операции. " +
